fix: added drizzle-kit to workspace root and reorganized seed and wipe scripts
