+++ packages/db/scripts/seed-database.ts
@@ -1,4 +1,5 @@
-import { eq, sql } from "drizzle-orm";
+import { eq, getTableColumns, sql } from "drizzle-orm";
+import { getTableName } from "drizzle-orm";
 
 import { db } from "../src/index";
 import { schema } from "../src/schemas/index";
@@ -263,6 +264,22 @@ await db.transaction(async (tx) => {
   );
 
   console.log("repair parts done");
+
+  await tx.execute(sql`SET CONSTRAINTS ALL DEFERRED;`);
+  for await (const table of Object.values(schema)) {
+    const tableName = getTableName(table);
+    const columns = getTableColumns(table);
+    if ("id" in columns) {
+      await tx.execute(
+        sql`SELECT pg_catalog.setval(pg_get_serial_sequence('${table}', 'id'), MAX(id)) FROM ${table};`,
+      );
+      console.log(`reset ${tableName}`);
+    } else {
+      console.log(`skipped ${tableName}`);
+    }
+  }
+
+  console.log("finished resetting sequences.");
 });
 
 console.log("done seeding");
